Move SpawnComponent to a dedicated module
