+++ vivarium/environment/components/eco_evo/__init__.py
@@ -1,1 +1,3 @@
-from .component import SpawnComponent, ConsumptionComponent, EnergyComponent, ReproductionComponent
+from .spawn.component import SpawnComponent
+from .spawn.controller import SpawnController
+from .component import ConsumptionComponent, EnergyComponent, ReproductionComponent
